requirements.txt
fastapi==0.108.0
ipython==8.5.0
keyboard==0.13.5
mangum==0.17.0
numpy==1.24.1
openai==1.6.1
Pillow==9.3.0
Pillow==10.1.0
protobuf==4.25.1
streamlit==1.29.0
streamlit_option_menu==0.3.6
transformers==4.36.2
webcolors==1.13
